1.修改了直方图统计方法，将统计的值改为百分比 (#64)
2,修改了测试直方图的方法

+++ ruoyi-admin/src/main/java/com/ruoyi/data/common/ObjectMap.java
@@ -16,9 +16,9 @@ public class ObjectMap {
     @Getter
     @Setter
     @JsonProperty("value")
-    private Integer value;
+    private Double value;
 
-    public ObjectMap(String key, Integer value) {
+    public ObjectMap(String key, Double value) {
         this.key = key;
         this.value = value;
     }
